pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compliance-ai-risk-platform"
version = "0.1.0"
description = "Compliance-Driven AI Risk Platform"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "click>=8.1",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
    "types-jsonschema>=4.20",
]

[tool.setuptools.packages.find]
include = ["orchestrator*"]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["scripts/"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "boto3"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "boto3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "botocore.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
